requirements.txt
# LLM Inference System Requirements
# Core ML/AI
vllm==0.3.0
torch>=2.1.0
transformers>=4.36.0
huggingface-hub

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Multimodal Support
pillow==10.2.0
python-multipart==0.0.6

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# AWS Integration
boto3==1.34.28

# Utilities
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.1.0
